pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pg-explain-viz"
version = "0.1.0"
description = "Visualize PostgreSQL EXPLAIN ANALYZE as ASCII tree"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "takawasi" }
]
keywords = ["postgresql", "explain", "visualization", "cli", "database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
pg-explain = "pg_explain.cli:main"

[project.urls]
Homepage = "https://github.com/takawasi/pg-explain-viz"
Repository = "https://github.com/takawasi/pg-explain-viz"

[tool.hatch.build.targets.wheel]
packages = ["pg_explain"]
